Improve saving of non-content files
Previously, when saving a non-content file, this would cause a rebuild
of all content-files. The `content/` directory would be walked for each
rebuild; on multiple saves in succession, the directory would be
concurrently walked, causing performance degredation and delay.

To fix this, the responsibility of processing the `content/` directory
(on file change) was given to a specific function. Since microtasks and
the microtasks that they themselves spawn (the "async Promise
callstack") are not cancellable, a new approach was needed. Things were
refactured to use a queue. And, the queue is emptied whenever a file is
changed.